+++ Response/WalletSend.cs
@@ -7,7 +7,7 @@ namespace TanDotNet.Response
 {
     public class WalletSend
     {
-        public int Balance { get; set; }
+        public ulong Balance { get; set; }
         public RedemptionMessage Message { get; set; }
     }
 }
